feat(cli): add radiologist-cli skeleton, remove per-package CLIs
Introduce the radiologist-cli workspace member as the single home for every
CLI command body (etl, core, registry, infer groups), stubbed with
NotImplementedError per the epic's interface contracts. The four business
packages (etl, core, registry, inference) become pure libraries: their
Typer/Hydra console-script entry points, cli.py modules, and CLI-only extras
are deleted so no two competing `radiologist` entry points can coexist.

Also adds radiologist.utils.cli with real output-format constants and
stubbed emit()/resolve_format()/exit_code_for(), rewires root and per-package
pyproject.toml dependency graphs to the new package, and updates workspace
tooling (conftest.py, Makefile, mkdocs.yml) to include radiologist-cli.

Closes #170
