+++ radiologist-cli/pyproject.toml
@@ -3,9 +3,9 @@ build-backend = "uv_build"
 requires = ["uv_build>=0.9.28,<0.10.0"]
 
 [project]
-name = "radiologist-utils"
+name = "radiologist-cli"
 version = "0.1.0"
-description = "Utility helpers for the radiologist pipeline"
+description = "Unified radiologist CLI dispatcher"
 readme = "README.md"
 license = "MIT"
 license-files = ["LICENSE"]
@@ -19,30 +19,21 @@ classifiers = [
 ]
 requires-python = ">=3.10"
 dependencies = [
-    "fsspec>=2026.4.0",
-    "numpy>=2.2.0",
-    "Pillow>=12.2.0",
-    "PyYAML>=6.0",
-    "rich>=15.0.0",
+    "radiologist-core[all]",
+    "radiologist-utils>=0.1.0",
+    "typer>=0.9.0",
+    "hydra-core>=1.3.0",
+    "omegaconf>=2.3.0",
 ]
 
 [project.optional-dependencies]
-all = [
-    "gcsfs>=2026.4.0",
-    "hydra-core>=1.3.0",
-    "lightning>=2.6.0",
-    "omegaconf>=2.3.0",
-    "torch>=2.12.0",
-    "wandb>=0.27.0"
-]
-gcs = ["gcsfs>=2026.4.0"]
-ml = [
-    "hydra-core>=1.3.0",
-    "lightning>=2.6.0",
-    "omegaconf>=2.3.0",
-    "torch>=2.12.0"
-    ]
-wandb = ["wandb>=0.27.0"]
+etl       = ["radiologist-etl[all]"]
+registry  = ["radiologist-registry[all]"]
+inference = ["radiologist-inference[all]"]
+all       = ["radiologist-etl[all]", "radiologist-registry[all]", "radiologist-inference[all]"]
+
+[project.scripts]
+radiologist = "radiologist.cli.main:main"
 
 [project.urls]
 Homepage = "https://github.com/CedrickArmel/radiologist"
@@ -54,11 +45,18 @@ Issues = "https://github.com/CedrickArmel/radiologist/issues"
 name = "cz_conventional_commits"
 version_provider = "uv"
 version_scheme = "pep440"
-tag_format = "${version}-radiologist-utils" # the component name can be a prefix or suffix with or without a separator
-ignored_tag_formats = ["${version}-radiologist-*"] # Avoid noise from other tags
+tag_format = "${version}-radiologist-cli"
+ignored_tag_formats = ["${version}-radiologist-*"]
 update_changelog_on_bump = true
-bump_message = "bump: radiologist-utils $current_version → $new_version"
+bump_message = "bump: radiologist-cli $current_version → $new_version"
 
 [tool.uv.build-backend]
 namespace = true
-module-name = "radiologist.utils"
+module-name = "radiologist.cli"
+
+[tool.uv.sources]
+radiologist-core = { workspace = true }
+radiologist-utils = { workspace = true }
+radiologist-etl = { workspace = true }
+radiologist-registry = { workspace = true }
+radiologist-inference = { workspace = true }
