+++ radiologist-core/pyproject.toml
@@ -33,9 +33,6 @@ classifiers = [
 ]
 requires-python = ">=3.10"
 
-[project.scripts]
-radiologist-core = "radiologist.core.train:main"
-
 [project.optional-dependencies]
 all = [
     "captum>=0.9.0",
