+++ radiologist-inference/pyproject.toml
@@ -31,10 +31,8 @@ all = [
     "fastapi>=0.100.0",
     "uvicorn>=0.23.0",
     "python-multipart>=0.0.9",
-    "typer>=0.9.0",
     "prometheus-client>=0.20.0",
 ]
-cli = ["typer>=0.9.0"]
 registry = ["wandb>=0.27.0"]
 serve = [
     "fastapi>=0.100.0",
@@ -42,9 +40,6 @@ serve = [
     "python-multipart>=0.0.9",
     "prometheus-client>=0.20.0",
 ]
-
-[project.scripts]
-radiologist = "radiologist.inference.cli:main"
 
 [project.urls]
 Homepage = "https://github.com/CedrickArmel/radiologist"
