+++ radiologist-inference/src/radiologist/inference/optional.py
@@ -33,11 +33,6 @@ except ImportError:
     _fastapi = None  # type: ignore[assignment]
 
 try:
-    import typer as _typer  # type: ignore[import-untyped]
-except ImportError:
-    _typer = None  # type: ignore[assignment]
-
-try:
     import uvicorn as _uvicorn  # type: ignore[import-untyped]
 except ImportError:
     _uvicorn = None  # type: ignore[assignment]
