+++ radiologist-registry/pyproject.toml
@@ -17,12 +17,8 @@ requires-python = ">=3.10"
 dependencies = []
 
 [project.optional-dependencies]
-all = ["typer>=0.9.0", "wandb>=0.27.0"]
-cli = ["typer>=0.9.0"]
+all = ["wandb>=0.27.0"]
 wandb = ["wandb>=0.27.0"]
-
-[project.scripts]
-radiologist-registry = "radiologist.registry.cli:main"
 
 [project.urls]
 Homepage = "https://github.com/CedrickArmel/radiologist"
